refactor lab offload CLI routing
